Reorganise tests in preparation for other PR
